multiple changes including experimental latent space flattening
draw.py: experimental version of reconstruct() that flattens
the latent space to be identical each iteration. There is
a corresponding sample_back() function to sample from this
space, and a sample_at() to provide the encoded values directly.

Note that this experiment (at this point) seems to not be working
right and will probably need retooling. But commiting now in its
current state.

sample.py now takes the number of rows/cols explicitly, as well as
an extra "flat" argument. This is also experimental, but the
intent is to be able to evenly sample the encoded space.

train-draw also has some further experiments, though note
that in the committed version the edge penality is still
disabled.